openharmony-schedule - 25

+++ openharmony/arkcompiler_ets_runtime/ecmascript/jspandafile/literal_data_extractor.cpp
@@ -354,7 +354,6 @@ JSHandle<JSFunction> LiteralDataExtractor::DefineMethodInLiteral(JSThread *threa
     if (module->IsSourceTextModule() && classKind == ClassKind::SENDABLE) {
         SourceTextModule::Cast(module->GetTaggedObject())->SetSendableEnv(thread, sendableEnv);
     }
-
     jsFunc->SetModule(thread, module.GetTaggedValue());
     jsFunc->SetLength(length);
     return jsFunc;
@@ -426,7 +425,7 @@ void LiteralDataExtractor::ExtractObjectDatas(JSThread *thread, const JSPandaFil
         bool flag = false;
         switch (tag) {
             case LiteralTag::INTEGER: {
-                jt = JSTaggedValue(static_cast<int32_t>(std::get<uint32_t>(value)));
+                jt = JSTaggedValue(std::get<uint32_t>(value));
                 break;
             }
             case LiteralTag::DOUBLE: {
